+++ src/orfimage.cpp
@@ -238,7 +238,7 @@ namespace Exiv2 {
     {
     }
 
-    bool OrfHeader::read(const byte* pData, uint32_t size)
+    bool OrfHeader::read(const byte* pData, uint64_t size)
     {
         if (size < 8) return false;
